# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: LLMOps Platform API
  description: |
    LLMOps平台API提供用户管理、项目管理、模型管理、推理服务、成本管理、监控日志、评测管理和知识库管理等核心功能。
    
    ## 认证方式
    - JWT Token认证
    - API Key认证
    - OAuth 2.0认证
    
    ## 限流策略
    - 用户限流: 1000 requests/hour
    - IP限流: 5000 requests/hour
    - API限流: 10000 requests/hour
    
    ## 错误处理
    所有API都返回统一的错误格式，包含错误码、错误信息和详细信息。
  version: 1.0.0
  contact:
    name: LLMOps Platform Support
    email: support@llmops.com
    url: https://llmops.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.llmops.com/v1
    description: 生产环境
  - url: https://api-test.llmops.com/v1
    description: 测试环境
  - url: https://api-dev.llmops.com/v1
    description: 开发环境
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /models:
    post:
      tags:
        - 模型管理
      summary: 注册模型
      description: 注册新模型
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRegistrationRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags:
        - 模型管理
      summary: 获取模型列表
      description: 获取模型列表，支持分页和过滤
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
  /models/{id}:
    get:
      tags:
        - 模型管理
      summary: 获取模型详情
      description: 获取指定模型的详细信息
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key认证
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      description: 搜索关键词
      schema:
        type: string
    SortParam:
      name: sort
      in: query
      description: 排序字段
      schema:
        type: string
        default: created_at:desc
...
    ModelIdParam:
      name: id
      in: path
      required: true
      description: 模型ID
      schema:
        type: integer
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    BaseResponse:
      type: object
      required:
        - code
        - message
        - data
        - meta
      properties:
        code:
          type: integer
          description: 响应状态码
        message:
          type: string
          description: 响应消息
        data:
          type: object
          description: 响应数据
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      required:
        - request_id
        - timestamp
        - version
      properties:
        request_id:
          type: string
          description: 请求ID
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
        version:
          type: string
          description: API版本
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - error
        - meta
      properties:
        code:
          type: integer
          description: 错误状态码
        message:
          type: string
          description: 错误消息
        error:
          $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorDetail:
      type: object
      required:
        - type
        - details
      properties:
        type:
          type: string
          description: 错误类型
        details:
          type: string
          description: 错误详情
        field:
          type: string
          description: 错误字段
        value:
          type: string
          description: 错误值

    # 分页响应
    PaginationResponse:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
        - has_next
        - has_prev
      properties:
        page:
          type: integer
          description: 当前页码
        per_page:
          type: integer
          description: 每页数量
        total:
          type: integer
          description: 总数量
        total_pages:
          type: integer
          description: 总页数
        has_next:
          type: boolean
          description: 是否有下一页
        has_prev:
          type: boolean
          description: 是否有上一页

    # 用户相关Schema
...
    ModelRegistrationRequest:
      type: object
      required:
        - name
        - code
        - type
        - category
        - framework
      properties:
        name:
          type: string
          description: 模型名称
          maxLength: 200
        code:
          type: string
          description: 模型代码
          maxLength: 50
        description:
          type: string
          description: 模型描述
        type:
          type: string
          enum: [llm, embedding, classification, generation]
          description: 模型类型
        category:
          type: string
          description: 模型分类
        framework:
          type: string
          description: 框架
        base_model:
          type: string
          description: 基础模型
        model_size:
          type: string
          description: 模型大小
        precision:
          type: string
          enum: [fp16, fp32, int8, int4]
          description: 精度
        tags:
          type: array
          items:
            type: string
          description: 标签
        metadata:
          type: object
          description: 元数据
        settings:
          type: object
          description: 模型设置
    ModelRegistrationResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                model:
                  $ref: '#/components/schemas/Model'
    ModelListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                models:
                  type: array
                  items:
                    $ref: '#/components/schemas/Model'
                pagination:
                  $ref: '#/components/schemas/PaginationResponse'
    ModelDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                model:
                  $ref: '#/components/schemas/Model'
...
    Model:
      type: object
      required:
        - id
        - uuid
        - name
        - code
        - type
        - category
        - framework
        - status
        - owner_id
        - project_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: 模型ID
        uuid:
          type: string
          format: uuid
          description: 模型UUID
        name:
          type: string
          description: 模型名称
        code:
          type: string
          description: 模型代码
        description:
          type: string
          description: 模型描述
        type:
          type: string
          enum: [llm, embedding, classification, generation]
          description: 模型类型
        category:
          type: string
          description: 模型分类
        framework:
          type: string
          description: 框架
        base_model:
          type: string
          description: 基础模型
        model_size:
          type: string
          description: 模型大小
        precision:
          type: string
          enum: [fp16, fp32, int8, int4]
          description: 精度
        status:
          type: string
          enum: [active, inactive, archived, deleted]
          description: 模型状态
        tags:
          type: array
          items:
            type: string
          description: 标签
        metadata:
          type: object
          description: 元数据
        settings:
          type: object
          description: 模型设置
        owner_id:
          type: integer
          description: 所有者ID
        project_id:
          type: integer
          description: 项目ID
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
...